refactor: split server/routes/universeBuilder.js into service-aligned sub-routers (#3392)
